+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Change Log
 
 ## Changed
+- compatible with ownCloud 11
 - moved admin settings to a dedicated ONLYOFFICE section in the sidebar
 
 ## 9.12.1